Style: board 메인/detail 보이는 정보  수정
