+++ src/components/features/board/BoardItem.jsx
@@ -1,20 +1,11 @@
 import React from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import { __deleteBoards } from "../../../redux/modules/boardSlice";
-import { Link, useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
 import styled from "styled-components";
 
 const BoardItem = ({ board }) => {
-  const dispatch = useDispatch();
   const { isLoading, error } = useSelector((state) => state.boardSlice);
-
-  const navigator = useNavigate();
-
-  // delete handler
-  const onDeleteHandler = (id) => {
-    dispatch(__deleteBoards(id));
-    navigator("/");
-  };
 
   if (isLoading) {
     return <div>loading...</div>;
@@ -25,15 +16,11 @@ const BoardItem = ({ board }) => {
 
   return (
     <div>
-      <div>
-        <DetailLink to={`/${board.id}`} key={board.id}>
-          자세히보기
-        </DetailLink>
+      <DetailLink to={`/${board.id}`} key={board.id}>
         <h3>{board.title}</h3>
         <p>{board.user}</p>
         <p>{board.content}</p>
-        <button onClick={() => onDeleteHandler(board.id)}>삭제</button>
-      </div>
+      </DetailLink>
     </div>
   );
 };
