+++ src/pages/BoardDetail.jsx
@@ -1,27 +1,35 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import { useSelector } from "react-redux";
+import { useSelector, useDispatch } from "react-redux";
 import Layout from "../components/UI/Layout";
-// import styled from "styled-components";
-//import { getBoardID } from "../redux/modules/boardSlice";
-import BoardItem from "../components/features/board/BoardItem";
+import { __deleteBoards } from "../redux/modules/boardSlice";
+import { useNavigate } from "react-router-dom";
+// import BoardItem from "../components/features/board/BoardItem";
 
 const Detail = () => {
-  //const dispatch = useDispatch();
+  const dispatch = useDispatch();
   const board = useSelector((state) => state.boardSlice.board);
-
+  const navigator = useNavigate();
   const param = useParams().id;
 
-  // useEffect(() => {
-  //   dispatch(getBoardID(id));
-  // }, [dispatch, id]);
+  // delete handler
+  const onDeleteHandler = (id) => {
+    dispatch(__deleteBoards(id));
+    navigator("/");
+  };
 
   return (
     <Layout>
       {board
         .filter((item) => item.id === param)
         .map((item) => {
-          return <BoardItem key={item.id} board={item} />;
+          return (
+            <div key={item.id}>
+              <div>{item.title}</div>
+              <div>{item.content}</div>
+              <button onClick={() => onDeleteHandler(item.id)}>삭제</button>
+            </div>
+          );
         })}
     </Layout>
   );
